+++ src/Webhook/Exception/WebhookParseFailedException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Torr\Storyblok\Webhook\Exception;
+
+use Torr\Storyblok\Exception\StoryblokException;
+
+final class WebhookParseFailedException extends \RuntimeException implements StoryblokException
+{
+}